PMA-127 - PackageTypePolicy

+++ app/Http/Controllers/PackageTypeController.php
@@ -9,6 +9,7 @@ use App\Http\Requests\PackageType\UpdateRequest;
 use App\Http\Resources\PackageTypeResource;
 use App\Http\Resources\PackageTypeResourceCollection;
 use App\Repositories\Contracts\PackageTypeRepository;
+use Illuminate\Auth\Access\AuthorizationException;
 
 class PackageTypeController extends Controller
 {
@@ -31,9 +32,12 @@ class PackageTypeController extends Controller
      *
      * @param IndexRequest $request
      * @return PackageTypeResourceCollection
+     * @throws AuthorizationException
      */
     public function index(IndexRequest $request)
     {
+        $this->authorize('list', [PackageType::class, $request->get('propertyId')]);
+
         $packageTypes = $this->packageTypeRepository->findBy($request->all());
 
         return new PackageTypeResourceCollection($packageTypes);
@@ -44,9 +48,12 @@ class PackageTypeController extends Controller
      *
      * @param  StoreRequest  $request
      * @return PackageTypeResource
+     * @throws AuthorizationException
      */
     public function store(StoreRequest $request)
     {
+        $this->authorize('store', [PackageType::class, $request->get('propertyId')]);
+
         $packageType = $this->packageTypeRepository->save($request->all());
 
         return new PackageTypeResource($packageType);
@@ -57,9 +64,12 @@ class PackageTypeController extends Controller
      *
      * @param PackageType $packageType
      * @return PackageTypeResource
+     * @throws AuthorizationException
      */
     public function show(PackageType $packageType)
     {
+        $this->authorize('show', $packageType);
+
         return new PackageTypeResource($packageType);
     }
 
@@ -69,9 +79,12 @@ class PackageTypeController extends Controller
      * @param UpdateRequest $request
      * @param PackageType $packageType
      * @return PackageTypeResource
+     * @throws AuthorizationException
      */
     public function update(UpdateRequest $request, PackageType $packageType)
     {
+        $this->authorize('update', $packageType);
+
         $packageType = $this->packageTypeRepository->update($packageType, $request->all());
 
         return new PackageTypeResource($packageType);
@@ -82,9 +95,12 @@ class PackageTypeController extends Controller
      *
      * @param PackageType $packageType
      * @return void
+     * @throws AuthorizationException
      */
     public function destroy(PackageType $packageType)
     {
+        $this->authorize('destroy', $packageType);
+
         $this->packageTypeRepository->delete($packageType);
 
         return response()->json(null, 204);
